Refactor paths and fix travis
