+++ lib/Helper.php
