+++ lib/Http/ResponseException.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * ownCloud - Richdocuments App
+ *
+ * @author Piotr Mrowczynski <piotr@owncloud.com>
+ * @copyright 2018 Piotr Mrowczynski <piotr@owncloud.com>
+ *
+ * This file is licensed under the Affero General Public License version 3 or
+ * later.
+ */
+namespace OCA\Richdocuments\Http;
+
+class ResponseException extends \Exception {
+	private $hint;
+
+	public function __construct($description, $hint = '') {
+		parent::__construct($description);
+		$this->hint = $hint;
+	}
+
+	public function getHint() {
+		return $this->hint;
+	}
+}
